backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlmodel>=0.0.16
aiofiles>=23.2.1
httpx>=0.27.0
trimesh>=4.3.0
pyrender>=0.1.45
watchdog>=4.0.0
python-multipart>=0.0.9
Pillow>=10.3.0
numpy>=1.26.0
pyglet>=2.0.0
